今後のベンチマーク追加予定に合わせてパッケージ構成を変更
