+++ src/jmh/kotlin/com/mapk/fastkfunction/fastkfunction/CallInstanceMethodBenchmark.kt
@@ -1,5 +1,6 @@
-package com.mapk.fastkfunction
+package com.mapk.fastkfunction.fastkfunction
 
+import com.mapk.fastkfunction.FastKFunction
 import com.mapk.fastkfunction.argumentbucket.ArgumentBucket
 import com.mapk.fastkfunction.benchmarktargets.Constructor5
 import org.openjdk.jmh.annotations.Benchmark
